requirements.txt
python-dotenv
wikipedia-api
openai
pandas
